change add friend notifications
