+++ src/pages/addFriends.tsx
@@ -37,16 +37,28 @@ function FriendCard({ username }: FriendCardProps) {
   const [submitting, setSubmitting] = useState(false);
   const { username: ownUsername } = useSelector(selectGlobalUser);
 
+  async function notifyFriend(username: string, friend: string) {
+    const URI = `/api/notifyFriend`;
+    const content = `${username} has added you as a friend!`;
+    try {
+      const response = await axios.post(URI, { content, friend });
+      console.log(response.data);
+    } catch (err) {
+      console.log(err);
+    }
+  }
+
   async function handleAddFriend(friend: string) {
     setSubmitting(true);
     try {
-      const response = await axios.post(URI, {
+      const addResponse = await axios.post(URI, {
         username: ownUsername,
         friend,
       });
-      console.log(response.data);
-      const { friends } = response.data;
+      console.log(addResponse.data);
+      const { friends } = addResponse.data;
       dispatch(setFriends(friends));
+      await notifyFriend(ownUsername, friend)
     } catch (err) {
       console.log(err);
     }
